Split browser specs for DEBT-370 Phase 2
