+++ app/(app)/app/practice/[sessionId]/components/practice-session-page-view-test-helpers.ts
@@ -1,0 +1,1 @@
+export const noop = () => undefined;
