+++ app/(app)/app/practice/[sessionId]/hooks/use-practice-session-page-controller-test-helpers.ts
@@ -1,0 +1,251 @@
+import { afterEach, beforeAll, beforeEach, expect } from 'vitest';
+import type { render } from 'vitest-browser-react';
+import type {
+  ActionErrorCode,
+  ActionResult,
+} from '@/src/adapters/controllers/action-result';
+import { ok } from '@/tests/test-helpers/ok';
+import {
+  createChoice,
+  createQuestionResponse,
+  createReviewResponse,
+  createReviewRow,
+} from './practice-session-page-controller.browser.fixtures';
+import {
+  getPracticeSessionPageControllerBrowserMocks,
+  resetPracticeSessionPageControllerBrowserMocks,
+} from './practice-session-page-controller.browser.setup';
+
+export let PracticeSessionPageControllerBookmarkPendingProbe: typeof import('./practice-session-page-controller.browser.probes').PracticeSessionPageControllerBookmarkPendingProbe;
+export let PracticeSessionPageControllerBookmarkProbe: typeof import('./practice-session-page-controller.browser.probes').PracticeSessionPageControllerBookmarkProbe;
+export let PracticeSessionPageControllerHookProbe: typeof import('./practice-session-page-controller.browser.probes').PracticeSessionPageControllerHookProbe;
+export let PracticeSessionPageControllerMarkForReviewProbe: typeof import('./practice-session-page-controller.browser.probes').PracticeSessionPageControllerMarkForReviewProbe;
+export let PracticeSessionPageControllerNavigationProbe: typeof import('./practice-session-page-controller.browser.probes').PracticeSessionPageControllerNavigationProbe;
+export let PracticeSessionPageControllerReviewProbe: typeof import('./practice-session-page-controller.browser.probes').PracticeSessionPageControllerReviewProbe;
+export let PracticeSessionPageControllerSummaryProbe: typeof import('./practice-session-page-controller.browser.probes').PracticeSessionPageControllerSummaryProbe;
+export let PracticeSessionPageControllerSubmitDuringReviewProbe: typeof import('./practice-session-page-controller.browser.probes').PracticeSessionPageControllerSubmitDuringReviewProbe;
+export let PracticeSessionPageControllerViewProbe: typeof import('./practice-session-page-controller.browser.probes').PracticeSessionPageControllerViewProbe;
+
+export const {
+  getNextQuestionMock,
+  submitAnswerMock,
+  getBookmarksMock,
+  toggleBookmarkMock,
+  getPracticeSessionReviewMock,
+  getCompletedSessionQuestionsWithFeedbackMock,
+  getPracticeSessionSummaryMock,
+  endPracticeSessionMock,
+  finalizeExamAnswersMock,
+  saveExamDraftAnswerMock,
+  setPracticeSessionQuestionMarkMock,
+} = getPracticeSessionPageControllerBrowserMocks();
+
+export const EMPTY_BOOKMARKS_RESULT = ok({ rows: [] });
+export const CHOICE_1 = createChoice({ id: 'choice_1' });
+export const CHOICE_2 = createChoice({
+  id: 'choice_2',
+  label: 'B',
+  textMd: 'Option B',
+  sortOrder: 2,
+});
+export const CHOICE_3 = createChoice({
+  id: 'choice_3',
+  label: 'C',
+  textMd: 'Option C',
+  sortOrder: 3,
+});
+
+export function errorResult(
+  code: ActionErrorCode,
+  message: string,
+): ActionResult<never> {
+  return {
+    ok: false,
+    error: { code, message },
+  };
+}
+
+export function mockBookmarksAndReview(
+  review: ReturnType<typeof createReviewResponse>,
+) {
+  getBookmarksMock.mockResolvedValue(EMPTY_BOOKMARKS_RESULT);
+  getPracticeSessionReviewMock.mockResolvedValue(ok(review));
+}
+
+export function mockExamReviewNavigationSession() {
+  getPracticeSessionSummaryMock.mockResolvedValue(
+    errorResult('CONFLICT', 'Practice session has not ended'),
+  );
+  mockBookmarksAndReview(
+    createReviewResponse({
+      mode: 'exam',
+      totalCount: 3,
+      answeredCount: 3,
+      markedCount: 0,
+      rows: [
+        createReviewRow({
+          questionId: 'question-1',
+          order: 1,
+          isAnswered: true,
+        }),
+        createReviewRow({
+          questionId: 'question-2',
+          order: 2,
+          isAnswered: true,
+        }),
+        createReviewRow({
+          questionId: 'question-3',
+          order: 3,
+          isAnswered: true,
+        }),
+      ],
+    }),
+  );
+  getNextQuestionMock.mockImplementation(async (input) => {
+    if (
+      typeof input === 'object' &&
+      input &&
+      'questionId' in input &&
+      typeof input.questionId === 'string'
+    ) {
+      const questionId = input.questionId;
+      const questionIndex =
+        questionId === 'question-1'
+          ? 0
+          : questionId === 'question-2'
+            ? 1
+            : questionId === 'question-3'
+              ? 2
+              : null;
+
+      if (questionIndex === null) {
+        throw new Error(`Unexpected questionId: ${questionId}`);
+      }
+
+      return ok(
+        createQuestionResponse({
+          questionId,
+          stemMd: `Stem ${questionId}`,
+          choices: [CHOICE_1, CHOICE_2, CHOICE_3],
+          session: {
+            mode: 'exam',
+            index: questionIndex,
+            total: 3,
+            isMarkedForReview: false,
+          },
+        }),
+      );
+    }
+
+    if (
+      typeof input === 'object' &&
+      input &&
+      'fromIndex' in input &&
+      typeof input.fromIndex === 'number'
+    ) {
+      return ok(null);
+    }
+
+    return ok(
+      createQuestionResponse({
+        questionId: 'question-3',
+        stemMd: 'Stem question-3',
+        choices: [CHOICE_1, CHOICE_2, CHOICE_3],
+        session: {
+          mode: 'exam',
+          index: 2,
+          total: 3,
+          isMarkedForReview: false,
+        },
+      }),
+    );
+  });
+}
+
+export async function openExamReviewQuestion(
+  screen: Awaited<ReturnType<typeof render>>,
+) {
+  await expect
+    .element(screen.getByTestId('question-id'))
+    .toHaveTextContent('question-3');
+  await screen.getByRole('button', { name: 'Review & Submit' }).click();
+  await expect
+    .element(screen.getByRole('heading', { name: 'Review & Submit' }))
+    .toBeVisible();
+  await screen
+    .getByRole('button', {
+      name: /Open question 2\..*Question 2.*Answered/i,
+    })
+    .click();
+  await expect
+    .element(screen.getByTestId('question-id'))
+    .toHaveTextContent('question-2');
+}
+
+export function setupPracticeSessionPageControllerBrowserSpec() {
+  beforeAll(async () => {
+    const probes = await import(
+      './practice-session-page-controller.browser.probes'
+    );
+    PracticeSessionPageControllerBookmarkPendingProbe =
+      probes.PracticeSessionPageControllerBookmarkPendingProbe;
+    PracticeSessionPageControllerBookmarkProbe =
+      probes.PracticeSessionPageControllerBookmarkProbe;
+    PracticeSessionPageControllerHookProbe =
+      probes.PracticeSessionPageControllerHookProbe;
+    PracticeSessionPageControllerMarkForReviewProbe =
+      probes.PracticeSessionPageControllerMarkForReviewProbe;
+    PracticeSessionPageControllerNavigationProbe =
+      probes.PracticeSessionPageControllerNavigationProbe;
+    PracticeSessionPageControllerReviewProbe =
+      probes.PracticeSessionPageControllerReviewProbe;
+    PracticeSessionPageControllerSummaryProbe =
+      probes.PracticeSessionPageControllerSummaryProbe;
+    PracticeSessionPageControllerSubmitDuringReviewProbe =
+      probes.PracticeSessionPageControllerSubmitDuringReviewProbe;
+    PracticeSessionPageControllerViewProbe =
+      probes.PracticeSessionPageControllerViewProbe;
+  });
+
+  beforeEach(() => {
+    getPracticeSessionSummaryMock.mockResolvedValue(
+      errorResult('CONFLICT', 'Practice session has not ended'),
+    );
+    getBookmarksMock.mockResolvedValue(EMPTY_BOOKMARKS_RESULT);
+    saveExamDraftAnswerMock.mockImplementation(async (input) =>
+      ok({
+        questionId:
+          typeof input === 'object' &&
+          input &&
+          'questionId' in input &&
+          typeof input.questionId === 'string'
+            ? input.questionId
+            : 'question-1',
+        markedForReview: false,
+        latestSelectedChoiceId: null,
+        latestIsCorrect: null,
+        latestAnsweredAt: null,
+        draftSelectedChoiceId:
+          typeof input === 'object' &&
+          input &&
+          'selectedChoiceId' in input &&
+          typeof input.selectedChoiceId === 'string'
+            ? input.selectedChoiceId
+            : 'choice_1',
+        draftSavedAt: new Date('2026-02-07T00:00:00.000Z'),
+        draftCumulativeMs:
+          typeof input === 'object' &&
+          input &&
+          'cumulativeMs' in input &&
+          typeof input.cumulativeMs === 'number'
+            ? input.cumulativeMs
+            : 1_000,
+      }),
+    );
+  });
+
+  afterEach(async () => {
+    await new Promise((resolve) => setTimeout(resolve, 0));
+    resetPracticeSessionPageControllerBrowserMocks();
+  });
+}
