+++ app/(app)/app/questions/[slug]/use-question-page-controller-test-helpers.tsx
@@ -1,0 +1,172 @@
+import { afterEach, beforeEach, vi } from 'vitest';
+import * as reportClientError from '@/lib/report-client-error';
+import type { QuestionOrigin } from '@/lib/routes';
+import * as bookmarkController from '@/src/adapters/controllers/bookmark-controller';
+import * as practiceController from '@/src/adapters/controllers/practice-controller';
+import * as questionController from '@/src/adapters/controllers/question-controller';
+import * as questionViewController from '@/src/adapters/controllers/question-view-controller';
+import type { GetBookmarksOutput } from '@/src/application/ports/bookmarks';
+import { ok } from '@/tests/test-helpers/ok';
+import { installReportClientErrorMocks } from '@/tests/test-helpers/report-client-error-mocks';
+import { useQuestionPageController } from './use-question-page-controller';
+
+vi.mock('@/src/adapters/controllers/question-view-controller', { spy: true });
+vi.mock('@/src/adapters/controllers/question-controller', { spy: true });
+vi.mock('@/src/adapters/controllers/practice-controller', { spy: true });
+vi.mock('@/src/adapters/controllers/bookmark-controller', { spy: true });
+vi.mock('@/lib/report-client-error', { spy: true });
+
+export const getQuestionBySlug = vi.mocked(
+  questionViewController.getQuestionBySlug,
+);
+export const getPreviousAttempt = vi.mocked(
+  questionViewController.getPreviousAttempt,
+);
+export const submitAnswer = vi.mocked(questionController.submitAnswer);
+export const getPracticeSessionReview = vi.mocked(
+  practiceController.getPracticeSessionReview,
+);
+export const getBookmarks = vi.mocked(bookmarkController.getBookmarks);
+export const toggleBookmark = vi.mocked(bookmarkController.toggleBookmark);
+export const reportClientErrorSpy = vi.mocked(
+  reportClientError.reportClientError,
+);
+
+installReportClientErrorMocks(reportClientError);
+
+const emptyBookmarksResult: { ok: true; data: GetBookmarksOutput } = ok({
+  rows: [],
+});
+
+export function setupQuestionPageControllerBrowserSpec() {
+  beforeEach(() => {
+    getBookmarks.mockResolvedValue(emptyBookmarksResult);
+    toggleBookmark.mockResolvedValue(ok({ bookmarked: false }));
+  });
+
+  afterEach(() => {
+    vi.resetAllMocks();
+  });
+}
+
+export function Probe({
+  slug = 'q-1',
+  mode,
+  sessionId,
+  attemptId,
+  from,
+  historySequence,
+  historyIndex,
+  onRender,
+}: {
+  slug?: string;
+  mode?: 'review' | null;
+  sessionId?: string;
+  attemptId?: string;
+  from?: QuestionOrigin | null;
+  historySequence?: readonly string[] | null;
+  historyIndex?: number | null;
+  onRender?: (snapshot: {
+    mode?: 'review' | null;
+    isLoadingPreviousAttempt: boolean;
+    reviewHydrationState: string | null;
+  }) => void;
+}) {
+  const output = useQuestionPageController({
+    slug,
+    mode,
+    sessionId,
+    attemptId,
+    from,
+    historySequence,
+    historyIndex,
+  });
+
+  onRender?.({
+    mode,
+    isLoadingPreviousAttempt: output.isLoadingPreviousAttempt,
+    reviewHydrationState: output.reviewHydrationState,
+  });
+
+  const total = output.sessionNavigation?.questions.length ?? null;
+  const index = output.sessionNavigation?.currentIndex ?? null;
+  const currentWasRetried =
+    index === null
+      ? null
+      : (output.sessionNavigation?.questions[index]?.wasRetried ?? null);
+  const prevSlug =
+    index === null || index <= 0
+      ? null
+      : (output.sessionNavigation?.questions[index - 1]?.slug ?? null);
+  const nextSlug =
+    index === null
+      ? null
+      : (output.sessionNavigation?.questions[index + 1]?.slug ?? null);
+
+  return (
+    <>
+      <div data-testid="load-status">{output.loadState.status}</div>
+      <div data-testid="question-slug">{output.question?.slug ?? ''}</div>
+      <div data-testid="selected-choice">{output.selectedChoiceId ?? ''}</div>
+      <div data-testid="attempt-id">{output.submitResult?.attemptId ?? ''}</div>
+      <div data-testid="unanswered-reveal-correct-choice">
+        {output.sessionUnansweredReveal?.correctChoiceId ?? ''}
+      </div>
+      <div data-testid="session-nav-total">{total ?? ''}</div>
+      <div data-testid="session-nav-index">{index ?? ''}</div>
+      <div data-testid="session-nav-current-was-retried">
+        {currentWasRetried === null ? '' : currentWasRetried ? 'true' : 'false'}
+      </div>
+      <div data-testid="session-nav-prev-slug">{prevSlug ?? ''}</div>
+      <div data-testid="session-nav-next-slug">{nextSlug ?? ''}</div>
+      <div data-testid="is-loading-previous-attempt">
+        {output.isLoadingPreviousAttempt ? 'true' : 'false'}
+      </div>
+      <div data-testid="review-hydration-state">
+        {output.reviewHydrationState ?? ''}
+      </div>
+      <div data-testid="bookmark-status">{output.bookmarkStatus}</div>
+      <div data-testid="is-bookmark-hydrated">
+        {output.isBookmarkHydrated ? 'true' : 'false'}
+      </div>
+      <div data-testid="is-bookmarked">
+        {output.isBookmarked ? 'true' : 'false'}
+      </div>
+      <button
+        type="button"
+        data-testid="select-choice-1"
+        onClick={() => output.onSelectChoice('choice-1')}
+      >
+        Select choice 1
+      </button>
+      <button
+        type="button"
+        data-testid="trigger-reattempt"
+        onClick={output.onReattempt}
+      >
+        Trigger reattempt
+      </button>
+      <button
+        type="button"
+        data-testid="trigger-submit"
+        onClick={() => void output.onSubmit()}
+      >
+        Trigger submit
+      </button>
+      <button
+        type="button"
+        data-testid="trigger-answer-as-new"
+        onClick={output.onAnswerAsNew}
+      >
+        Trigger answer as new
+      </button>
+      <button
+        type="button"
+        data-testid="trigger-toggle-bookmark"
+        onClick={() => void output.onToggleBookmark()}
+      >
+        Trigger toggle bookmark
+      </button>
+    </>
+  );
+}
